chore(frontend/deps): bump 5 minor/patch deps + clean up lint warnings
Subset of dependabot #20 (js-minor-and-patch group), excluding
eslint-plugin-react-refresh@0.5.2 because it requires eslint >=9 and
the project is still on eslint 8.

Bumps:
- react-router-dom 7.12.0 -> 7.14.2
- @types/prismjs 1.26.5 -> 1.26.6
- @typescript-eslint/eslint-plugin 8.56.1 -> 8.59.0
- autoprefixer 10.4.22 -> 10.5.0
- postcss 8.5.10 -> 8.5.12

Also clears 4 pre-existing lint warnings that were silently failing PR
CI (`--max-warnings 0`):
- FloatingChatPanel: pin the language-only effect with an explicit
  eslint-disable + rationale
- LanguageProvider: split storage helpers into languageStorage.ts so
  the provider file only exports a component (fixes
  react-refresh/only-export-components)
