+++ frontend/package.json
@@ -15,22 +15,22 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-markdown": "^10.1.0",
-    "react-router-dom": "^7.11.0",
+    "react-router-dom": "7.14.2",
     "react-simple-code-editor": "^0.14.1",
     "remark-gfm": "^4.0.1"
   },
   "devDependencies": {
-    "@types/prismjs": "^1.26.5",
+    "@types/prismjs": "1.26.6",
     "@types/react": "^18.2.43",
     "@types/react-dom": "^18.2.17",
-    "@typescript-eslint/eslint-plugin": "^8.56.1",
+    "@typescript-eslint/eslint-plugin": "8.59.0",
     "@typescript-eslint/parser": "^8.56.1",
     "@vitejs/plugin-react": "^4.2.1",
-    "autoprefixer": "^10.4.16",
+    "autoprefixer": "10.5.0",
     "eslint": "^8.55.0",
     "eslint-plugin-react-hooks": "^4.6.0",
     "eslint-plugin-react-refresh": "^0.4.5",
-    "postcss": "^8.5.10",
+    "postcss": "8.5.12",
     "tailwindcss": "^3.4.0",
     "typescript": "^5.2.2",
     "vite": "^7.3.2"
